+++ src/models/__tests__/Question.spec.ts
@@ -171,7 +171,6 @@ describe('update', () => {
 
 describe('create', () => {
   const originalInsert = Question.insert;
-  const originalFindByQuestionText = Question.findByQuestionText;
   let insertQuery;
   let question;
 
@@ -193,7 +192,6 @@ describe('create', () => {
   afterEach(() => {
     // jest.resetAllMocks();
     Question.insert = originalInsert;
-    Question.findByQuestionText = originalFindByQuestionText;
   });
 
   it('returns the Question without errors if it is valid', async () => {
@@ -223,28 +221,12 @@ describe('create', () => {
     expect(response.errors[0]).toBe('Question text can\'t be blank');
   });
 
-  it('returns the Question with an error if the question already exists', async () => {
-    const mockFindBy = jest.fn();
-    (Question.findByQuestionText as jest.Mock) = mockFindBy;
-    mockFindBy.mockResolvedValueOnce(question);
-
-    const result = await question.create(context);
-    expect(mockFindBy).toHaveBeenCalledTimes(1);
-    expect(result.errors.length).toBe(1);
-    expect(result.errors[0]).toEqual('Question with this question text already exists');
-  });
-
   it('returns the newly added Question', async () => {
-    const mockFindBy = jest.fn();
-    (Question.findByQuestionText as jest.Mock) = mockFindBy;
-    mockFindBy.mockResolvedValueOnce(null);
-
     const mockFindById = jest.fn();
     (Question.findById as jest.Mock) = mockFindById;
     mockFindById.mockResolvedValueOnce(question);
 
     const result = await question.create(context);
-    expect(mockFindBy).toHaveBeenCalledTimes(1);
     expect(mockFindById).toHaveBeenCalledTimes(1);
     expect(insertQuery).toHaveBeenCalledTimes(1);
     expect(result.errors.length).toBe(0);
@@ -290,50 +272,5 @@ describe('delete', () => {
     const result = await question.delete(context);
     expect(result.errors.length).toBe(0);
     expect(result).toEqual(question);
-  });
-});
-
-describe('findByQuestionText', () => {
-  const originalQuery = Question.query;
-
-  let localQuery;
-  let context;
-  let question;
-
-  beforeEach(() => {
-    jest.resetAllMocks();
-
-    localQuery = jest.fn();
-    (Question.query as jest.Mock) = localQuery;
-
-    context = buildContext(logger, mockToken());
-
-    question = new Question({
-      templateId: casual.integer(1, 999),
-      sectionId: casual.integer(1, 999),
-      id: casual.integer(1, 9),
-      questionText: casual.sentences(5),
-      displayOrder: casual.integer(1, 9),
-    })
-  });
-
-  afterEach(() => {
-    jest.clearAllMocks();
-    Question.query = originalQuery;
-  });
-
-  it('should call query with correct params and return the question', async () => {
-    localQuery.mockResolvedValueOnce([question]);
-    const result = await Question.findByQuestionText('Question query', context, question.questionText, question.sectionId, question.templateId);
-    const expectedSql = 'SELECT * FROM questions WHERE LOWER(questionText) = ? AND sectionId = ? AND templateId = ?';
-    expect(localQuery).toHaveBeenCalledTimes(1);
-    expect(localQuery).toHaveBeenLastCalledWith(context, expectedSql, [question.questionText.toLowerCase(), question.sectionId.toString(), question.templateId.toString()], 'Question query')
-    expect(result).toEqual(question);
-  });
-
-  it('should return null if it finds no Question', async () => {
-    localQuery.mockResolvedValueOnce([]);
-    const result = await Question.findByQuestionText('Question query', context, question.questionText, question.sectionId, question.templateId);
-    expect(result).toEqual(null);
   });
 });